Create dict_autoupdate in Bug class
dict_autoupdate is a parameter that can be passed in Bug's constructor
that instructs Bug class to auto update local dictionary when a
set/add/update method is called.

When this parameter is False, the Bug class will work as always: if you
perform a set/add/update action then read the attribute related to the
performed action, it will not show the new value, but the old one.

When dict_autoupdate is True, the Bug class will update internal
dictionary using the new value passed in set/add/update method.

To do that some changes have to be done:
 - Create helper methods inside of Bug class
 - Modify set/add/update methods adding calls for updating dictionary
 - Add a new optional parameter to Bug's constructor
 - Modify Bugzilla class to add support to dict_autoupdate parameter
 - Update test_query2.txt to avoid tests to be broken by new parameter

Also done:
 - A new unit test for this feature
 - Pylint fixes